services/feature-002-ai-predictor/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.0
google-generativeai==0.3.0
supabase==2.3.4
pytest==7.4.2
pytest-flask==1.2.0
